Make a new lunch roulette session maximizing the gossip to exchange

Make a new lunch roulette session maximizing the gossip to exchange

rsocialroulette(current_frame, past_sessions, m, algorithm = c("mdgp", "srs"))

Arguments

current_frame
A tibble containing the participants of the current round, i.e. it has a column `id` containing a unique identifier and a `date` column representing the date of the session.

past_sessions
A list of partitionLists, i.e. each partition is a list of vectors containing the id of the members of the corresponding group.

m
minimum group size, i.e. all groups will be at least size m.

Value

A partitionList containing the partitioning of current_frame maximizing the overall sum of gossip to be exchanged.

Examples

library(socialroulette)
#Distribute 5 ppl into groups of at least 2
frame <- tibble(id=sprintf("id%.3d@math.su.se", 1:5), date=Sys.Date())

#> Error in tibble(id = sprintf("id%.3d@math.su.se", 1:5), date = Sys.Date()): could not find function "tibble"

rsocialroulette(frame, past_sessions=NULL, m=2, algorithm="srs")

#> Error in rsocialroulette(frame, past_sessions = NULL, m = 2, algorithm = "srs"): "id" %in% colnames(current_frame) is not TRUE

#Same, but to test that algorithm="mdgp" also works with past_session=NULL
rsocialroulette(frame, past_sessions=NULL, m=2, algorithm="mdgp")

#> Error in rsocialroulette(frame, past_sessions = NULL, m = 2, algorithm = "mdgp"): "id" %in% colnames(current_frame) is not TRUE

# Create a history of previous sessions as well as the current frame from a population of 100
ppl <- tibble(id=sprintf("id%.3d@math.su.se", 1:100))

#> Error in tibble(id = sprintf("id%.3d@math.su.se", 1:100)): could not find function "tibble"

session_dates <- seq(as.Date("2021-04-01"), length.out=4, by="1 week")
# Simulate changing participation each week (with prob 0.7 they will attend)
frames <- map_df( session_dates, ~ ppl %>% slice_sample(n = rbinom(1,nrow(.), prob=0.7)) %>% mutate(date=.x))

#> Error in map_df(session_dates, ~ppl %>% slice_sample(n = rbinom(1, nrow(.),     prob = 0.7)) %>% mutate(date = .x)): could not find function "map_df"

# Frame for the current session is the last one
current_frame <- frames %>% filter(date == max(date))

#> Error in frames %>% filter(date == max(date)): could not find function "%>%"

# Simulate the partitions for the last 3 weeks
past_sessions <- frames %>% filter(date < max(session_dates)) %>%
  group_split(date) %>%
  map(~rsocialroulette(current_frame=.x, past_sessions=NULL, m=4)) %>%
  setNames(session_dates[1:3])

#> Error in frames %>% filter(date < max(session_dates)) %>% group_split(date) %>%     map(~rsocialroulette(current_frame = .x, past_sessions = NULL,         m = 4)) %>% setNames(session_dates[1:3]): could not find function "%>%"

# Make the partition, which tries to bring together those who have not met in a long time
rsocialroulette(current_frame, past_session, m=4, algorithm="mdgp")

#> Error in is.data.frame(x): object 'current_frame' not found