Convert session specification to Lai and Hao (2016) MDGP solver input specification

Convert session specification to Lai and Hao (2016) MDGP solver input specification

write_mdgp_specfile(current_frame, past_sessions, m)

Arguments

current_frame
A pair-distance tibble. The frame needs to contain a column denoted `id`

past_sessions
A list of sessions. Each sessions consists of a session specification in list format.

m
The minimum group size, i.e. all groups have size at least m

Value

File name of the generated specification file

Examples

 ppl <- tibble(id=sprintf("id%.3d", 1:100))

#> Error in tibble(id = sprintf("id%.3d", 1:100)): could not find function "tibble"

 session_dates <- seq(as.Date("2021-04-01"), length.out=4, by="1 week")
 frames <- map_df( session_dates, ~ ppl %>% slice_sample(n = rbinom(1,nrow(ppl), prob=0.7)) %>% mutate(date=.x))

#> Error in map_df(session_dates, ~ppl %>% slice_sample(n = rbinom(1, nrow(ppl),     prob = 0.7)) %>% mutate(date = .x)): could not find function "map_df"

 past_sessions <- frames %>% group_split(date) %>% map(~sample_groups(.x, m=4)) %>% setNames(session_dates) %>% .[1:3]

#> Error in frames %>% group_split(date) %>% map(~sample_groups(.x, m = 4)) %>%     setNames(session_dates) %>% .[1:3]: could not find function "%>%"

 current_frame <- frames %>% filter(date == max(date))

#> Error in frames %>% filter(date == max(date)): could not find function "%>%"

 spec_file <- write_mdgp_specfile(current_frame, past_sessions, m=4)

#> Error in write_mdgp_specfile(current_frame, past_sessions, m = 4): could not find function "write_mdgp_specfile"

 mdgp_solver(spec_file)

#> Error in mdgp_solver(spec_file): could not find function "mdgp_solver"