Convert partition specification to Lai and Hao (2016) MDGP solver input specification.

Generate a tempfile() containing the appropriate specification formed from current_frame and past_sessions.

write_mdgp_specfile(current_frame, past_sessions, m)

Arguments

current_frame
A pair-distance tibble. The frame needs to contain a column denoted `id`

past_sessions
A list of sessions. Each sessions consists of a session specification in list format.

m
The minimum group size, i.e. all groups have size at least m

Value

File name of the generated specification file

Details

The specification format has the following header line $$n\> m\> ds\> n_1\> n_1\> \ldots n_G \> n_G$$ Here, each group size \(n_g, g=1,\ldots,G\) is repeated twice, because the solver has the opportunity to specify a lower as well as an upper boundary for each group size. The keyword `ds` for the solver in the above means that the groups can be of different sizes as determined by the respective \(n_g\).

Examples

frame  <- tibble::tibble(id=sprintf("id%.2d", 1:5), date=as.Date("2021-04-28"))
round1 <- list(c("id02", "id03", "id04"), c("id01", "id05"))
spec_file <- socialroulette:::write_mdgp_specfile(frame, past_sessions=list("2021-04-21"=round1), m=2)
cat(stringr::str_c(readLines(spec_file), collapse="\n"))

#> 5 2 ds 3 3 2 2
#> 0 1 14
#> 0 2 14
#> 0 3 14
#> 0 4 7
#> 1 2 7
#> 1 3 7
#> 1 4 14
#> 2 3 7
#> 2 4 14
#> 3 4 14